+++ core/src/main/kotlin/com/r3corda/core/testing/InMemoryVaultService.kt
@@ -4,8 +4,8 @@ import com.r3corda.core.ThreadBox
 import com.r3corda.core.contracts.*
 import com.r3corda.core.crypto.SecureHash
 import com.r3corda.core.node.ServiceHub
-import com.r3corda.core.node.services.Wallet
-import com.r3corda.core.node.services.WalletService
+import com.r3corda.core.node.services.Vault
+import com.r3corda.core.node.services.VaultService
 import com.r3corda.core.serialization.SingletonSerializeAsToken
 import com.r3corda.core.transactions.WireTransaction
 import com.r3corda.core.utilities.loggerFor
@@ -17,88 +17,88 @@ import java.util.*
 import javax.annotation.concurrent.ThreadSafe
 
 /**
- * This class implements a simple, in memory wallet that tracks states that are owned by us, and also has a convenience
- * method to auto-generate some self-issued cash states that can be used for test trading. A real wallet would persist
- * states relevant to us into a database and once such a wallet is implemented, this scaffolding can be removed.
+ * This class implements a simple, in memory vault that tracks states that are owned by us, and also has a convenience
+ * method to auto-generate some self-issued cash states that can be used for test trading. A real vault would persist
+ * states relevant to us into a database and once such a vault is implemented, this scaffolding can be removed.
  */
 @ThreadSafe
-open class InMemoryWalletService(protected val services: ServiceHub) : SingletonSerializeAsToken(), WalletService {
-    open protected val log = loggerFor<InMemoryWalletService>()
+open class InMemoryVaultService(protected val services: ServiceHub) : SingletonSerializeAsToken(), VaultService {
+    open protected val log = loggerFor<InMemoryVaultService>()
 
     // Variables inside InnerState are protected with a lock by the ThreadBox and aren't in scope unless you're
     // inside mutex.locked {} code block. So we can't forget to take the lock unless we accidentally leak a reference
-    // to wallet somewhere.
+    // to vault somewhere.
     protected class InnerState {
-        var wallet = Wallet(emptyList<StateAndRef<ContractState>>())
+        var vault = Vault(emptyList<StateAndRef<ContractState>>())
     }
 
     protected val mutex = ThreadBox(InnerState())
 
-    override val currentWallet: Wallet get() = mutex.locked { wallet }
+    override val currentVault: Vault get() = mutex.locked { vault }
 
-    private val _updatesPublisher = PublishSubject.create<Wallet.Update>()
+    private val _updatesPublisher = PublishSubject.create<Vault.Update>()
 
-    override val updates: Observable<Wallet.Update>
+    override val updates: Observable<Vault.Update>
         get() = _updatesPublisher
 
     /**
      * Returns a snapshot of the heads of LinearStates.
      */
     override val linearHeads: Map<UniqueIdentifier, StateAndRef<LinearState>>
-        get() = currentWallet.let { wallet ->
-            wallet.states.filterStatesOfType<LinearState>().associateBy { it.state.data.linearId }.mapValues { it.value }
+        get() = currentVault.let { vault ->
+            vault.states.filterStatesOfType<LinearState>().associateBy { it.state.data.linearId }.mapValues { it.value }
         }
 
-    override fun notifyAll(txns: Iterable<WireTransaction>): Wallet {
+    override fun notifyAll(txns: Iterable<WireTransaction>): Vault {
         val ourKeys = services.keyManagementService.keys.keys
 
         // Note how terribly incomplete this all is!
         //
         // - We don't notify anyone of anything, there are no event listeners.
-        // - We don't handle or even notice invalidations due to double spends of things in our wallet.
+        // - We don't handle or even notice invalidations due to double spends of things in our vault.
         // - We have no concept of confidence (for txns where there is no definite finality).
         // - No notification that keys are used, for the case where we observe a spend of our own states.
         // - No ability to create complex spends.
-        // - No logging or tracking of how the wallet got into this state.
+        // - No logging or tracking of how the vault got into this state.
         // - No persistence.
         // - Does tx relevancy calculation and key management need to be interlocked? Probably yes.
         //
         // ... and many other things .... (Wallet.java in bitcoinj is several thousand lines long)
 
-        var netDelta = Wallet.NoUpdate
-        val changedWallet = mutex.locked {
-            // Starting from the current wallet, keep applying the transaction updates, calculating a new Wallet each
+        var netDelta = Vault.NoUpdate
+        val changedVault = mutex.locked {
+            // Starting from the current vault, keep applying the transaction updates, calculating a new vault each
             // time, until we get to the result (this is perhaps a bit inefficient, but it's functional and easily
             // unit tested).
-            val walletAndNetDelta = txns.fold(Pair(currentWallet, Wallet.NoUpdate)) { walletAndDelta, tx ->
-                val (wallet, delta) = walletAndDelta.first.update(tx, ourKeys)
-                val combinedDelta = delta + walletAndDelta.second
-                Pair(wallet, combinedDelta)
+            val vaultAndNetDelta = txns.fold(Pair(currentVault, Vault.NoUpdate)) { vaultAndDelta, tx ->
+                val (vault, delta) = vaultAndDelta.first.update(tx, ourKeys)
+                val combinedDelta = delta + vaultAndDelta.second
+                Pair(vault, combinedDelta)
             }
 
-            wallet = walletAndNetDelta.first
-            netDelta = walletAndNetDelta.second
-            return@locked wallet
+            vault = vaultAndNetDelta.first
+            netDelta = vaultAndNetDelta.second
+            return@locked vault
         }
 
-        if (netDelta != Wallet.NoUpdate) {
+        if (netDelta != Vault.NoUpdate) {
             _updatesPublisher.onNext(netDelta)
         }
-        return changedWallet
+        return changedVault
     }
 
     private fun isRelevant(state: ContractState, ourKeys: Set<PublicKey>): Boolean {
         return if (state is OwnableState) {
             state.owner in ourKeys
         } else if (state is LinearState) {
-            // It's potentially of interest to the wallet
+            // It's potentially of interest to the vault
             state.isRelevant(ourKeys)
         } else {
             false
         }
     }
 
-    private fun Wallet.update(tx: WireTransaction, ourKeys: Set<PublicKey>): Pair<Wallet, Wallet.Update> {
+    private fun Vault.update(tx: WireTransaction, ourKeys: Set<PublicKey>): Pair<Vault, Vault.Update> {
         val ourNewStates = tx.outputs.
                 filter { isRelevant(it.data, ourKeys) }.
                 map { tx.outRef<ContractState>(it.data) }
@@ -108,19 +108,19 @@ open class InMemoryWalletService(protected val services: ServiceHub) : Singleton
 
         // Is transaction irrelevant?
         if (consumed.isEmpty() && ourNewStates.isEmpty()) {
-            log.trace { "tx ${tx.id} was irrelevant to this wallet, ignoring" }
-            return Pair(this, Wallet.NoUpdate)
+            log.trace { "tx ${tx.id} was irrelevant to this vault, ignoring" }
+            return Pair(this, Vault.NoUpdate)
         }
 
-        val change = Wallet.Update(consumed, HashSet(ourNewStates))
+        val change = Vault.Update(consumed, HashSet(ourNewStates))
 
-        // And calculate the new wallet.
+        // And calculate the new vault.
         val newStates = states.filter { it.ref !in consumed } + ourNewStates
 
         log.trace {
-            "Applied tx ${tx.id.prefixChars()} to the wallet: consumed ${consumed.size} states and added ${newStates.size}"
+            "Applied tx ${tx.id.prefixChars()} to the vault: consumed ${consumed.size} states and added ${newStates.size}"
         }
 
-        return Pair(Wallet(newStates), change)
+        return Pair(Vault(newStates), change)
     }
 }