+++ node/src/test/kotlin/com/r3corda/node/services/NodeVaultServiceTest.kt
@@ -3,11 +3,11 @@ package com.r3corda.node.services
 import com.r3corda.contracts.testing.fillWithSomeTestCash
 import com.r3corda.core.contracts.DOLLARS
 import com.r3corda.core.node.services.TxWritableStorageService
-import com.r3corda.core.node.services.WalletService
+import com.r3corda.core.node.services.VaultService
 import com.r3corda.core.transactions.SignedTransaction
 import com.r3corda.core.utilities.DUMMY_NOTARY
 import com.r3corda.core.utilities.LogHelper
-import com.r3corda.node.services.wallet.NodeWalletService
+import com.r3corda.node.services.vault.NodeVaultService
 import com.r3corda.node.utilities.configureDatabase
 import com.r3corda.node.utilities.databaseTransaction
 import com.r3corda.testing.node.MockServices
@@ -19,42 +19,42 @@ import org.junit.Test
 import java.io.Closeable
 import java.util.*
 
-class NodeWalletServiceTest {
+class NodeVaultServiceTest {
     lateinit var dataSource: Closeable
 
     @Before
     fun setUp() {
-        LogHelper.setLevel(NodeWalletService::class)
+        LogHelper.setLevel(NodeVaultService::class)
         dataSource = configureDatabase(makeTestDataSourceProperties()).first
     }
 
     @After
     fun tearDown() {
         dataSource.close()
-        LogHelper.reset(NodeWalletService::class)
+        LogHelper.reset(NodeVaultService::class)
     }
 
     @Test
     fun `states not local to instance`() {
         databaseTransaction {
             val services1 = object : MockServices() {
-                override val walletService: WalletService = NodeWalletService(this)
+                override val vaultService: VaultService = NodeVaultService(this)
 
                 override fun recordTransactions(txs: Iterable<SignedTransaction>) {
                     for (stx in txs) {
                         storageService.validatedTransactions.addTransaction(stx)
-                        walletService.notify(stx.tx)
+                        vaultService.notify(stx.tx)
                     }
                 }
             }
             services1.fillWithSomeTestCash(100.DOLLARS, DUMMY_NOTARY, 3, 3, Random(0L))
 
-            val w1 = services1.walletService.currentWallet
+            val w1 = services1.vaultService.currentVault
             assertThat(w1.states).hasSize(3)
 
             val originalStorage = services1.storageService
             val services2 = object : MockServices() {
-                override val walletService: WalletService = NodeWalletService(this)
+                override val vaultService: VaultService = NodeVaultService(this)
 
                 // We need to be able to find the same transactions as before, too.
                 override val storageService: TxWritableStorageService get() = originalStorage
@@ -62,12 +62,12 @@ class NodeWalletServiceTest {
                 override fun recordTransactions(txs: Iterable<SignedTransaction>) {
                     for (stx in txs) {
                         storageService.validatedTransactions.addTransaction(stx)
-                        walletService.notify(stx.tx)
+                        vaultService.notify(stx.tx)
                     }
                 }
             }
 
-            val w2 = services2.walletService.currentWallet
+            val w2 = services2.vaultService.currentVault
             assertThat(w2.states).hasSize(3)
         }
     }